+++ cmd/notificationsslack.go
@@ -107,15 +107,17 @@ var deleteProjectSlackNotificationCmd = &cobra.Command{
 			cmd.Help()
 			os.Exit(1)
 		}
-		deleteResult, err := pClient.DeleteSlackNotificationFromProject(notificationFlags.Project, notificationFlags.NotificationName)
-		handleError(err)
-		var addedProject api.NotificationSlack
-		err = json.Unmarshal([]byte(deleteResult), &addedProject)
-		handleError(err)
-		resultData := output.Result{
-			Result: "success",
+		if yesNo(fmt.Sprintf("You are attempting to delete notification '%s' from project '%s', are you sure?", notificationFlags.NotificationName, notificationFlags.Project)) {
+			deleteResult, err := pClient.DeleteSlackNotificationFromProject(notificationFlags.Project, notificationFlags.NotificationName)
+			handleError(err)
+			var addedProject api.NotificationSlack
+			err = json.Unmarshal([]byte(deleteResult), &addedProject)
+			handleError(err)
+			resultData := output.Result{
+				Result: "success",
+			}
+			output.RenderResult(resultData, outputOptions)
 		}
-		output.RenderResult(resultData, outputOptions)
 	},
 }
 
@@ -130,12 +132,16 @@ var deleteSlackNotificationCmd = &cobra.Command{
 			cmd.Help()
 			os.Exit(1)
 		}
-		deleteResult, err := pClient.DeleteSlackNotification(notificationFlags.NotificationName)
-		handleError(err)
-		resultData := output.Result{
-			Result: string(deleteResult),
+		fmt.Println(fmt.Sprintf("Deleting notification %s", notificationFlags.NotificationName))
+
+		if yesNo(fmt.Sprintf("You are attempting to delete notification '%s' from lagoon, are you sure?", notificationFlags.NotificationName)) {
+			deleteResult, err := pClient.DeleteSlackNotification(notificationFlags.NotificationName)
+			handleError(err)
+			resultData := output.Result{
+				Result: string(deleteResult),
+			}
+			output.RenderResult(resultData, outputOptions)
 		}
-		output.RenderResult(resultData, outputOptions)
 	},
 }
